Make sure the temporary file in join_aug_pred() is always unique
Previous implementation could result in accidentally using (appending to)
the temporary file used in a previous function call (when --nocleanup option
was used and the join_aug_pred() was called more than 2 times). Such situation
occurred for instance when the --augustus_ab_initio was used in the ETP mode
and it led to erroneous behavior.